More resource safety tests

+++ src/main/scala/fs2/Process1Ops.scala
@@ -11,7 +11,7 @@ trait Process1Ops[+F[_],+O] { self: Stream[F,O] =>
   def last: Stream[F,Option[O]] = self pipe process1.last
 
   /** Alias for `self pipe [[process1.take]](n)`. */
-  def take(n: Int): Stream[F,O] = self pipe process1.take(n)
+  def take(n: Long): Stream[F,O] = self pipe process1.take(n)
 
   /** Alias for `self pipe [[process1.unchunk]]`. */
   def unchunk: Stream[F,O] = self pipe process1.unchunk
